Finish launchpad analyzer cleanup
