+++ aurora_dashboard.html
@@ -13,7 +13,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link rel="stylesheet" href="static/css/site.css">
-    <script src="static/js/site-launchpad.js" defer></script>
+    <script src="static/js/site-launchpad.js" type="module"></script>
 </head>
 <body data-page="dashboard">
     <div class="site-shell">
