+++ index.html
@@ -17,7 +17,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link rel="stylesheet" href="static/css/site.css">
-    <script src="static/js/site-launchpad.js" defer></script>
+    <script src="static/js/site-launchpad.js" type="module"></script>
 </head>
 <body data-page="launchpad">
     <div class="site-shell">
